refactor: separate under construction content
